Test: add service facade and poro tests for resturant and yelp

+++ spec/poros/current_weather_spec.rb
@@ -14,7 +14,7 @@ RSpec.describe CurrentWeather, :vcr do
     expect(current.temperature).to be_a Float
     expect(current.feels_like).to be_a Float
     expect(current.humidity).to be_a Integer
-    expect(current.uvi).to be_a Integer
+    expect(current.uvi).to be_a Float
     expect(current.visibility).to be_a Integer
     expect(current.conditions).to be_a String
     expect(current.icon).to be_a String
